minutes_agent/requirements.txt
langgraph
langchain_cohere
langchain_core
python-dotenv
langchain_community
tenacity